add diferent JS files
