+++ book.js
@@ -1,0 +1,13 @@
+// Declare the Book class
+class Book {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
+}
+
+export default Book;
+
+// Declare the Display Class
+
+// Preserve data in a browser using LocalStorage
