+++ index.html
@@ -26,6 +26,6 @@
     </section>
     </main>
 
-<script src="./main.js"></script>    
+<script type="module" src="./app.js"></script>    
 </body>
 </html>
